Fix backup removal command to check target_dir
Ensure target_dir is set before removing backup.

+++ src/backup.sh
@@ -650,7 +650,8 @@ apply_retention_policy() {
         fi
 
         log_msg INFO "Removing expired backup: ${target_dir%/}/$name"
-        rm -rf -- "${target_dir%/}/$name"
+        target_dir=${target_dir%/}
+        rm -rf -- "${target_dir:?}/$name"
     done
 }
